+++ tests/test_blood.py
@@ -9,20 +9,10 @@ class _TestBlood(_TestSituation4Mc12):
         cls.pre_mes = '== Blood trails in this server =='
 
     def assert_bleeding(self, yes, no):
-        if yes == 0:
-            yes = set()
-        if no == 0:
-            no = set()
+        self.assert_property(yes, no, 'C', lambda c: c.is_bleeding)
 
-        if yes == 1:
-            yes = {client for client in self.server.client_manager.clients if client not in no}
-        if no == 1:
-            no = {client for client in self.server.client_manager.clients if client not in yes}
-
-        for client in yes:
-            self.assertTrue(client.is_bleeding, client)
-        for client in no:
-            self.assertFalse(client.is_bleeding, client)
+    def assert_smeared(self, yes, no):
+        self.assert_property(yes, no, 'A', lambda a: a.blood_smeared)
 
 class TestBlood_01_Basic(_TestBlood):
     def test_01_wrongarguments(self):
@@ -479,7 +469,7 @@ class TestBlood_05_Clean(_TestBlood):
         self.c0.assert_no_ooc()
         self.c1.assert_ooc('{} cleaned the blood trail in area {}.'
                            .format(self.c2.name, self.a4_name), over=True)
-        self.c2.assert_ooc('You cleaned the blood trail in the current area.'
+        self.c2.assert_ooc('You cleaned the blood trail in your area.'
                            .format(self.a4_name), over=True)
         self.c3.assert_ooc('The blood trail in your area was cleaned.'
                            .format(self.c3_cname, self.a4_name), over=True)
@@ -538,9 +528,9 @@ class TestBlood_05_Clean(_TestBlood):
 
         self.c1.ooc('/bloodtrail_clean {}, {}'
                     .format(self.a0_name, self.area6.id))
-        self.c1.assert_ooc('You cleaned the blood trails in areas {}, {}.'
+        self.c1.assert_ooc('You cleaned the blood trails in areas {} and {}.'
                            .format(self.a0_name, self.a6_name), over=True)
-        self.c2.assert_ooc('{} cleaned the blood trails in areas {}, {}.'
+        self.c2.assert_ooc('{} cleaned the blood trails in areas {} and {}.'
                            .format(self.c1.name, self.a0_name, self.a6_name), over=True)
 
         self.c1.ooc('/bloodtrail_list')
@@ -578,3 +568,241 @@ class TestBlood_05_Clean(_TestBlood):
         self.c1.ooc('/bloodtrail_list')
         self.c1.assert_ooc(self.mes, over=True)
 
+class TestBlood_06_Smear(_TestBlood):
+    @classmethod
+    def setUpClass(cls):
+        """
+        Situation:
+                 -------------
+                |            |
+                A0 -> A4 -> A5
+               |
+               ---> A6 -> A7(*)
+        """
+
+        super().setUpClass()
+        cls.c1.ooc('/bloodtrail 0')
+        cls.c0.move_area(4)
+        cls.c0.move_area(5)
+        cls.c0.move_area(0)
+        cls.c0.move_area(6)
+        cls.c0.move_area(7)
+
+        cls.c0.discard_all()
+        cls.c1.discard_all()
+        cls.c2.discard_all()
+        cls.c3.discard_all()
+
+        cls.c1.ooc('/bloodtrail_list')
+        cls.mes = ('{}'
+                   '\r\n*({}) {}: {}, {}, {}'
+                   '\r\n*({}) {}: {}, {}'
+                   '\r\n*({}) {}: {}, {}'
+                   '\r\n*({}) {}: {}, {}'
+                   '\r\n*({}) {}: {}'
+                   .format(cls.pre_mes,
+                           cls.area0.id, cls.a0_name, cls.a4_name, cls.a5_name, cls.a6_name,
+                           cls.area4.id, cls.a4_name, cls.a0_name, cls.a5_name,
+                           cls.area5.id, cls.a5_name, cls.a0_name, cls.a4_name,
+                           cls.area6.id, cls.a6_name, cls.a0_name, cls.a7_name,
+                           cls.area7.id, cls.a7_name, cls.a6_name))
+        cls.c1.assert_ooc(cls.mes, over=True)
+
+    def test_01_wrongarguments(self):
+        """
+        Situation: Clients try to use /bloodtrail_clean incorrectly.
+        """
+
+        # Non-staff clean other areas
+        self.c3.ooc('/bloodtrail_smear 7')
+        self.c0.assert_no_ooc()
+        self.c1.assert_no_ooc()
+        self.c2.assert_no_ooc()
+        self.c3.assert_ooc('You must be authorized to do that.', over=True)
+        self.c1.ooc('/bloodtrail_list')
+        self.c1.assert_ooc(self.mes, over=True)
+
+        # Invalid area names
+        self.c2.ooc('/bloodtrail_smear Not an area name')
+        self.c0.assert_no_ooc()
+        self.c1.assert_no_ooc()
+        self.c2.assert_ooc('Could not parse area Not an area name', over=True)
+        self.c3.assert_no_ooc()
+        self.c2.ooc('/bloodtrail_list')
+        self.c2.assert_ooc(self.mes, over=True)
+
+        self.c2.ooc('/bloodtrail_smear 100') # No area called 100, or with ID 100 in test scenario
+        self.c0.assert_no_ooc()
+        self.c1.assert_no_ooc()
+        self.c2.assert_ooc('Could not parse area 100', over=True)
+        self.c3.assert_no_ooc()
+        self.c1.ooc('/bloodtrail_list')
+        self.c1.assert_ooc(self.mes, over=True)
+
+        # Not using ,\ for areas with , in their names ("Class Trial Room,\ 2")
+        self.c2.ooc('/bloodtrail_smear Class Trial Room, 2, Test 4')
+        self.c0.assert_no_ooc()
+        self.c1.assert_no_ooc()
+        self.c2.assert_ooc('Could not parse area Class Trial Room', over=True)
+        self.c3.assert_no_ooc()
+        self.c1.ooc('/bloodtrail_list')
+        self.c1.assert_ooc(self.mes, over=True)
+
+    def test_02_smearownarea(self):
+        """
+        Situation: C2 smears their own area using no arguments. C1 verifies smear later.
+        Note that this doesn't affect existing trails leading to the area, but just the blood
+        inside the area.
+        """
+
+        self.c2.ooc('/bloodtrail_smear')
+        self.c0.assert_no_ooc()
+        self.c1.assert_ooc('{} smeared the blood trail in area {}.'
+                           .format(self.c2.name, self.a4_name), over=True)
+        self.c2.assert_ooc('You smeared the blood trail in your area.'
+                           .format(self.a4_name), over=True)
+        self.c3.assert_ooc('The blood trail in your area was smeared.'
+                           .format(self.c3_cname, self.a4_name), over=True)
+
+        self.c1.ooc('/bloodtrail_list')
+        self.mes = ('{}'
+                   '\r\n*({}) {}: {}, {}, {}'
+                   '\r\n*({}) {}: {}, {} {}'
+                   '\r\n*({}) {}: {}, {}'
+                   '\r\n*({}) {}: {}, {}'
+                   '\r\n*({}) {}: {}'
+                   .format(self.pre_mes,
+                           self.area0.id, self.a0_name, self.a4_name, self.a5_name, self.a6_name,
+                           self.area4.id, self.a4_name, self.a0_name, self.a5_name, '(SMEARED)',
+                           self.area5.id, self.a5_name, self.a0_name, self.a4_name,
+                           self.area6.id, self.a6_name, self.a0_name, self.a7_name,
+                           self.area7.id, self.a7_name, self.a6_name))
+        self.c1.assert_ooc(self.mes, over=True)
+
+    def test_03_smearfail(self):
+        """
+        Situation: C2 attempts to smear smeared areas and non-staff C0 attempts to smear clean area.
+        These two fail.
+        """
+
+        self.c2.ooc('/bloodtrail_smear')
+        self.c0.assert_no_ooc()
+        self.c1.assert_no_ooc()
+        self.c2.assert_ooc('Area {} already has its blood trails smeared.'.format(self.a4_name),
+                           over=True)
+        self.c3.assert_no_ooc()
+
+        self.c1.ooc('/bloodtrail_list')
+        self.mes = ('{}'
+                   '\r\n*({}) {}: {}, {}, {}'
+                   '\r\n*({}) {}: {}, {} {}'
+                   '\r\n*({}) {}: {}, {}'
+                   '\r\n*({}) {}: {}, {}'
+                   '\r\n*({}) {}: {}'
+                   .format(self.pre_mes,
+                           self.area0.id, self.a0_name, self.a4_name, self.a5_name, self.a6_name,
+                           self.area4.id, self.a4_name, self.a0_name, self.a5_name, '(SMEARED)',
+                           self.area5.id, self.a5_name, self.a0_name, self.a4_name,
+                           self.area6.id, self.a6_name, self.a0_name, self.a7_name,
+                           self.area7.id, self.a7_name, self.a6_name))
+        self.c1.assert_ooc(self.mes, over=True)
+
+        self.c3.move_area(2)
+        self.c3.ooc('/bloodtrail_smear')
+        self.c0.assert_no_ooc()
+        self.c1.assert_no_ooc()
+        self.c2.assert_no_ooc()
+        self.c3.assert_ooc('There is no blood in area {}.'.format(self.a2_name), over=True)
+
+        self.c1.ooc('/bloodtrail_list')
+        self.c1.assert_ooc(self.mes, over=True)
+
+        self.c3.move_area(4)
+
+    def test_04_smearotherareas(self):
+        """
+        Situation: C1 smears multiple areas simultaneously.
+        Assumption: a0_name < a6_name
+        """
+
+        self.c1.ooc('/bloodtrail_smear {}, {}'
+                    .format(self.a0_name, self.area6.id))
+        self.c1.assert_ooc('You smeared the blood trails in areas {} and {}.'
+                           .format(self.a0_name, self.a6_name), over=True)
+        self.c2.assert_ooc('{} smeared the blood trails in areas {} and {}.'
+                           .format(self.c1.name, self.a0_name, self.a6_name), over=True)
+
+        self.c1.ooc('/bloodtrail_list')
+        self.mes = ('{}'
+                   '\r\n*({}) {}: {}, {}, {} {}'
+                   '\r\n*({}) {}: {}, {} {}'
+                   '\r\n*({}) {}: {}, {}'
+                   '\r\n*({}) {}: {}, {} {}'
+                   '\r\n*({}) {}: {}'
+                   .format(self.pre_mes,
+                           0, self.a0_name, self.a4_name, self.a5_name, self.a6_name, '(SMEARED)',
+                           self.area4.id, self.a4_name, self.a0_name, self.a5_name, '(SMEARED)',
+                           self.area5.id, self.a5_name, self.a0_name, self.a4_name,
+                           self.area6.id, self.a6_name, self.a0_name, self.a7_name, '(SMEARED)',
+                           self.area7.id, self.a7_name, self.a6_name))
+        self.c1.assert_ooc(self.mes, over=True)
+
+    def test_05_smearinbleeding(self):
+        """
+        Situation: C0 and C1 attempt to clean the area where bleeding C0 is. This succeeds... once.
+        """
+
+        self.mes = ('{}'
+                   '\r\n*({}) {}: {}, {}, {} {}'
+                   '\r\n*({}) {}: {}, {} {}'
+                   '\r\n*({}) {}: {}, {}'
+                   '\r\n*({}) {}: {}, {} {}'
+                   '\r\n*({}) {}: {} {}'
+                   .format(self.pre_mes,
+                           0, self.a0_name, self.a4_name, self.a5_name, self.a6_name, '(SMEARED)',
+                           self.area4.id, self.a4_name, self.a0_name, self.a5_name, '(SMEARED)',
+                           self.area5.id, self.a5_name, self.a0_name, self.a4_name,
+                           self.area6.id, self.a6_name, self.a0_name, self.a7_name, '(SMEARED)',
+                           self.area7.id, self.a7_name, self.a6_name, '(SMEARED)'))
+
+        self.c0.ooc('/bloodtrail_smear')
+        self.c0.assert_ooc('You smeared the blood trail in your area.',
+                           over=True)
+        self.c1.assert_ooc('{} smeared the blood trail in area {}.'
+                           .format(self.c0_cname, self.a7_name), over=True)
+        self.c2.assert_ooc('{} smeared the blood trail in area {}.'
+                           .format(self.c0_cname, self.a7_name), over=True)
+
+        self.c1.ooc('/bloodtrail_list')
+        self.c1.assert_ooc(self.mes, over=True)
+
+        self.c1.ooc('/bloodtrail_smear {}'.format(self.a7_name))
+        self.c1.assert_ooc('Area {} already has its blood trails smeared.'
+                           .format(self.a7_name), over=True)
+
+        self.c1.ooc('/bloodtrail_list')
+        self.c1.assert_ooc(self.mes, over=True)
+
+    def test_06_cleansmeared(self):
+        """
+        Situation: C2 and C0 attempt to clean area 6 and 7 respectively. C2 succeeds, C0 doesn't as
+        they are still bleeding.
+        """
+
+        self.mes = ('{}'
+                   '\r\n*({}) {}: {}, {}, {} {}'
+                   '\r\n*({}) {}: {}, {} {}'
+                   '\r\n*({}) {}: {}, {}'
+                   '\r\n*({}) {}: {} {}'
+                   .format(self.pre_mes,
+                           0, self.a0_name, self.a4_name, self.a5_name, self.a6_name, '(SMEARED)',
+                           self.area4.id, self.a4_name, self.a0_name, self.a5_name, '(SMEARED)',
+                           self.area5.id, self.a5_name, self.a0_name, self.a4_name,
+                           self.area7.id, self.a7_name, self.a6_name, '(SMEARED)'))
+
+        self.c1.ooc('/bloodtrail_clean {}'.format(self.a6_name))
+        self.c1.assert_ooc('You cleaned the blood trail in area {}.'
+                           .format(self.a6_name), over=True)
+        self.c2.assert_ooc('{} cleaned the blood trail in area {}.'
+                           .format(self.c1.name, self.a6_name), over=True)
+        self.assertFalse(self.area6.blood_smeared)